feat(dashboard): move nav links to a left sidebar, keep profile in the header (#7)
