+++ src/components/layout/DashboardShell.tsx
@@ -1,15 +1,16 @@
 'use client';
 
 import DashboardHeader from './DashboardHeader';
+import DashboardSidebar from './DashboardSidebar';
 
-/**
- * Coquille du tableau de bord : header épuré (sans sidebar), contenu centré.
- */
 export default function DashboardShell({ children }: { children: React.ReactNode }) {
   return (
     <div className="min-h-screen flex flex-col bg-gaming-dark">
       <DashboardHeader />
-      <main className="flex-1">{children}</main>
+      <div className="flex flex-1">
+        <DashboardSidebar />
+        <main className="flex-1 min-w-0">{children}</main>
+      </div>
     </div>
   );
 }
